feat(views): migrate Auth pages (Login/SignUp) to shared @repo/views (M7)
Move LoginPage and SignUpPage (incl. their *Content components) from
apps/app into packages/views. Authentication logic stays in the app and
is injected through a new AuthProvider/useAuth seam, so @repo/views holds
no better-auth (or other client-specific auth) references.

- Add packages/views/src/auth: AuthProvider/useAuth context exposing
  useSession + signInWithEmail/signUpWithEmail/signInWithGitHub/
  signInWithGoogle, typed against neverthrow (already a views dep).
- Register ./LoginPage, ./SignUpPage and ./auth subpath exports.
- apps/app: add integrations/auth.ts (webAuth from better-auth-client)
  and wrap children with AuthProvider inside RefineProvider, next to the
  existing PlatformProvider. Web login/sign-up routes import from
  @repo/views.
- Desktop runs in local mode with no auth flow (no better-auth client,
  no login/signup routes); none are added, matching the task's explicit
  "only if desktop has its own auth flow" condition.

Verified: apps/app `bun run compile` clean, desktop `vite build` ok.
