+++ apps/app/src/routes/sign-up.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { SignUpPage } from "@/pages/SignUpPage";
+import { SignUpPage } from "@repo/views/SignUpPage";
 
 export const Route = createFileRoute("/sign-up")({
   head: () => ({
