+++ packages/views/package.json
@@ -25,12 +25,15 @@
     "./AgentsPage": "./src/pages/AgentsPage/index.ts",
     "./AgentDetailPage": "./src/pages/AgentDetailPage/index.ts",
     "./SkillsPage": "./src/pages/SkillsPage/index.ts",
+    "./LoginPage": "./src/pages/LoginPage/index.ts",
+    "./SignUpPage": "./src/pages/SignUpPage/index.ts",
     "./PipelinesPage": "./src/pages/PipelinesPage/index.ts",
     "./PipelineDetailPage": "./src/pages/PipelineDetailPage/index.ts",
     "./PageHeader": "./src/components/PageHeader/index.ts",
     "./PageLoadingState": "./src/components/PageLoadingState/index.ts",
     "./ToastContainer": "./src/components/ToastContainer.tsx",
     "./platform": "./src/platform/index.ts",
+    "./auth": "./src/auth/index.ts",
     "./constants": "./src/constants.ts",
     "./store/sidebarStore": "./src/store/sidebarStore/index.ts",
     "./store/toastStore": "./src/store/toastStore/index.ts"
